+++ src/main/migrations/index.ts
@@ -1,5 +1,6 @@
 import { Migration } from '../../common/types';
+import migration024 from './001_add_system_name_to_templates';
 
-const migrations: Migration[] = [];
+const migrations: Migration[] = [migration024];
 
 export default migrations;