+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added at [0.2.4]
 
 - **Project Migration:** Implemented a project migration mechanism to automatically update project structure to new versions.
+- **Templates:** Added `export_name` to templates for robust import/export.
 
 ### Changed at [0.2.4]